app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sqlalchemy
cachetools
python-dotenv
email-validator
pydantic

